undo binary tree + update SeedsManager

+++ src/ISeedsManager.sol
@@ -2,6 +2,6 @@
 pragma solidity ^0.8.13;
 
 interface ISeedsManager {
-    function getSeed(uint256) external view returns (uint16);
-    function addSeeds(uint16[] calldata) external;
+    function getSeed(uint32) external view returns (uint32);
+    function addSeeds(uint32[] calldata) external;
 }
